refactored dark-light mode toggling
